backend/requirements/base.txt
Django>=5.2,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
celery>=5.3
redis>=5.0
Pillow>=10.0
django-cors-headers>=4.3
django-filter>=23.5
django-auditlog>=2.3
drf-spectacular>=0.27
qrcode[pil]>=7.4
scikit-learn>=1.3
imbalanced-learn>=0.11
shap>=0.43
joblib>=1.3
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
